Add debops.keyring Python3 module to generate keyring documentation
Features:

* Generate RST documentation from the machine readable files in this
  repository.
* Run extensive tests to ensure that the keyring is consistent on CI
  tests.

When this is merged I can add debops-keyring to debops/docs. The
Makefile will be called when `docs/_prepare` and the Makefile is present
in repositories.  This allows submodules to hook into the docs build.

I also prepared to include the OpenPGP fingerprints into the rendered
docs but this is currently disabled.
"TODO: Disabled because the docs are currently served via non-encrypted and legacy HTTP so showing the public key fingerprints does not make sense."